refactor(api): move key admin routes
