+++ app/admin/financeiro/saldo/page.tsx
@@ -39,19 +39,17 @@ export default function SaldoPage() {
     const fetchData = async () => {
       try {
         setLoading(true)
-        const saldoRes = await fetch('/admin/api/asaas/saldo')
+        const saldoRes = await fetch('/api/asaas/saldo')
         if (saldoRes.ok) {
           const data: { balance: number } = await saldoRes.json()
           setSaldoDisponivel(data.balance)
         }
-        const statsRes = await fetch(
-          '/admin/api/asaas/estatisticas?status=PENDING',
-        )
+        const statsRes = await fetch('/api/asaas/estatisticas?status=PENDING')
         if (statsRes.ok) {
           const stats: Statistics = await statsRes.json()
           setALiberar(stats.netValue)
         }
-        const extratoRes = await fetch('/admin/api/asaas/extrato')
+        const extratoRes = await fetch('/api/asaas/extrato')
         if (extratoRes.ok) {
           const data = await extratoRes.json()
           setExtrato(data.data || [])
