+++ app/api/lider/[id]/route.ts
